fix: suggestion popover
