+++ packages/core/src/editor/components/show-popover.tsx
@@ -31,9 +31,7 @@ function _ShowPopover(props: ShowPopoverProps) {
       query: showIfKey || '',
       from: 'variable',
       editor,
-    })
-      .map((variable) => variable.name)
-      .slice(0, 5);
+    }).map((variable) => variable.name);
   }, [variables, showIfKey, editor]);
 
   const isValidWhenKey = showIfKey || autoCompleteOptions.includes(showIfKey);
