+++ packages/core/src/editor/nodes/variable/variable-suggestions.tsx
@@ -3,93 +3,40 @@ import {
   DEFAULT_VARIABLES,
   Variables,
 } from '@/editor/provider';
-import { cn } from '@/editor/utils/classname';
 import { processVariables } from '@/editor/utils/variable';
 import { ReactRenderer } from '@tiptap/react';
 import { SuggestionOptions } from '@tiptap/suggestion';
-import {
-  ArrowDown,
-  ArrowUp,
-  Braces,
-  CornerDownLeft,
-  Variable,
-} from 'lucide-react';
-import {
-  forwardRef,
-  useEffect,
-  useImperativeHandle,
-  useState,
-  useRef,
-} from 'react';
+import { forwardRef, useImperativeHandle, useState } from 'react';
 import tippy, { GetReferenceClientRect } from 'tippy.js';
+import { VariablePopover, VariablePopoverRef } from './variable-popover';
+import { useRef } from 'react';
 
 export const VariableList = forwardRef((props: any, ref) => {
   const { items = [] } = props;
 
-  const [selectedIndex, setSelectedIndex] = useState(0);
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
-  const itemRefs = useRef<(HTMLButtonElement | null)[]>([]);
-
-  const selectItem = (index: number) => {
-    const item = props.items[index];
-    if (!item) {
-      return;
-    }
-
-    props.command({ id: item });
-  };
-
-  const scrollSelectedIntoView = (index: number) => {
-    const container = scrollContainerRef.current;
-    const selectedItem = itemRefs.current[index];
-
-    if (!container || !selectedItem) {
-      return;
-    }
-
-    const containerRect = container.getBoundingClientRect();
-    const itemRect = selectedItem.getBoundingClientRect();
-
-    const padding = 4;
-    if (itemRect.bottom > containerRect.bottom) {
-      // Scroll down if item is below viewport
-      container.scrollTop += itemRect.bottom - containerRect.bottom + padding;
-    } else if (itemRect.top < containerRect.top) {
-      // Scroll up if item is above viewport
-      container.scrollTop += itemRect.top - containerRect.top - padding;
-    }
-  };
-
-  useEffect(() => {
-    setSelectedIndex(0);
-    // Reset scroll position when items change
-    if (scrollContainerRef.current) {
-      scrollContainerRef.current.scrollTop = 0;
-    }
-    // Reset item refs array
-    itemRefs.current = props.items.map(() => null);
-  }, [props.items]);
-
-  useEffect(() => {
-    scrollSelectedIntoView(selectedIndex);
-  }, [selectedIndex]);
+  const popoverRef = useRef<VariablePopoverRef>(null);
 
   useImperativeHandle(ref, () => ({
     onKeyDown: ({ event }: { event: KeyboardEvent }) => {
+      if (!popoverRef.current) {
+        return false;
+      }
+
+      const { moveUp, moveDown, select } = popoverRef.current || {};
       if (event.key === 'ArrowUp') {
         event.preventDefault();
-        setSelectedIndex((selectedIndex + items.length - 1) % items.length);
+        moveUp();
         return true;
       }
 
       if (event.key === 'ArrowDown') {
         event.preventDefault();
-        setSelectedIndex((selectedIndex + 1) % items.length);
+        moveDown();
         return true;
       }
 
       if (event.key === 'Enter') {
-        selectItem(selectedIndex);
+        select();
         return true;
       }
 
@@ -98,81 +45,18 @@ export const VariableList = forwardRef((props: any, ref) => {
   }));
 
   return (
-    <div className="mly-z-50 mly-w-64 mly-rounded-lg mly-border mly-border-gray-200 mly-bg-white mly-shadow-md mly-transition-all">
-      <div className="mly-flex mly-items-center mly-justify-between mly-gap-2 mly-border-b mly-border-gray-200 mly-bg-soft-gray/40 mly-px-1 mly-py-1.5 mly-text-gray-500">
-        <span className="mly-text-xs mly-uppercase">Variables</span>
-        <VariableIcon>
-          <Braces className="mly-size-3 mly-stroke-[2.5]" />
-        </VariableIcon>
-      </div>
-
-      <div
-        ref={scrollContainerRef}
-        className="mly-max-h-52 mly-overflow-y-auto mly-scrollbar-thin mly-scrollbar-track-transparent mly-scrollbar-thumb-gray-200"
-      >
-        <div className="mly-flex mly-flex-col mly-gap-0.5 mly-p-1">
-          {items?.length ? (
-            items?.map((item: string, index: number) => (
-              <button
-                key={index}
-                ref={(el) => (itemRefs.current[index] = el)}
-                onClick={() => selectItem(index)}
-                className={cn(
-                  'mly-flex mly-w-full mly-items-center mly-gap-2 mly-rounded-md mly-px-2 mly-py-1 mly-text-left mly-font-mono mly-text-sm mly-text-gray-900 hover:mly-bg-soft-gray',
-                  index === selectedIndex ? 'mly-bg-soft-gray' : 'mly-bg-white'
-                )}
-              >
-                <Braces className="mly-size-3 mly-stroke-[2.5] mly-text-rose-600" />
-                {item}
-              </button>
-            ))
-          ) : (
-            <div className="mly-flex mly-h-7 mly-w-full mly-items-center mly-gap-2 mly-rounded-md mly-px-2 mly-py-1 mly-text-left mly-font-mono mly-text-[13px] mly-text-gray-900 hover:mly-bg-soft-gray">
-              No result
-            </div>
-          )}
-        </div>
-      </div>
-
-      <div className="mly-flex mly-items-center mly-justify-between mly-gap-2 mly-border-t mly-border-gray-200 mly-px-1 mly-py-1.5 mly-text-gray-500">
-        <div className="mly-flex mly-items-center mly-gap-1">
-          <VariableIcon>
-            <ArrowDown className="mly-size-3 mly-stroke-[2.5]" />
-          </VariableIcon>
-          <VariableIcon>
-            <ArrowUp className="mly-size-3 mly-stroke-[2.5]" />
-          </VariableIcon>
-          <span className="mly-text-xs mly-text-gray-500">Navigate</span>
-        </div>
-        <VariableIcon>
-          <CornerDownLeft className="mly-size-3 mly-stroke-[2.5]" />
-        </VariableIcon>
-      </div>
-    </div>
+    <VariablePopover
+      items={items.map((item: string) => ({
+        name: item,
+        required: false,
+      }))}
+      onSelectItem={(value) => {
+        props.command({ id: value });
+      }}
+      ref={popoverRef}
+    />
   );
 });
-
-VariableList.displayName = 'VariableList';
-
-type VariableIconProps = {
-  className?: string;
-  children: React.ReactNode;
-};
-
-function VariableIcon(props: VariableIconProps) {
-  const { className, children } = props;
-
-  return (
-    <div
-      className={cn(
-        'mly-flex mly-size-5 mly-items-center mly-justify-center mly-rounded-md mly-border',
-        className
-      )}
-    >
-      {children}
-    </div>
-  );
-}
 
 VariableList.displayName = 'VariableList';
